+++ vscode/src/completions/latency.ts
@@ -8,7 +8,19 @@ export const defaultLatency = {
 }
 
 // Languages with lower performance get additional latency to avoid spamming users with unhelpful suggestions
-const lowPerformanceLanguageIds = new Set(['css', 'html', 'scss', 'vue', 'dart', 'json', 'yaml', 'postcss'])
+export const lowPerformanceLanguageIds = new Set([
+    'css',
+    'html',
+    'scss',
+    'vue',
+    'dart',
+    'json',
+    'yaml',
+    'postcss',
+    'markdown',
+    'plaintext',
+    'xml',
+])
 
 let userMetrics = {
     sessionTimestamp: 0,
@@ -21,10 +33,12 @@ let userMetrics = {
 // Start when the last 5 suggestions were not accepted
 // Increment latency by 200ms linearly up to max latency
 // Reset every 5 minutes, or on file change, or on accepting a suggestion
-export function getLatency(provider: string, fsPath: string, languageId?: string): number {
+export function getLatency(provider: string, fsPath: string, languageId?: string, nodeType?: string): number {
+    // set base latency based on provider and low performance languages or comments when available
     let baseline = provider === 'anthropic' ? 0 : defaultLatency.baseline
-    // set base latency based on provider and low performance languages
-    if (!languageId || (languageId && lowPerformanceLanguageIds.has(languageId))) {
+    const isLowPerformance = languageId && lowPerformanceLanguageIds.has(languageId)
+    const isComment = nodeType === 'comment'
+    if (!languageId || isLowPerformance || isComment) {
         baseline = defaultLatency.lowPerformance
     }
 
@@ -42,15 +56,10 @@ export function getLatency(provider: string, fsPath: string, languageId?: string
     userMetrics.suggested++
     userMetrics.fsPath = fsPath
 
-    // Start after 5 rejected suggestions
-    if (userMetrics.suggested < 5) {
-        return baseline
-    }
-
     const total = Math.max(baseline, Math.min(baseline + userMetrics.currentLatency, defaultLatency.max))
 
-    // Increase latency linearly up to max
-    if (userMetrics.currentLatency < defaultLatency.max) {
+    // Increase latency linearly up to max after 5 rejected suggestions
+    if (userMetrics.suggested >= 5 && userMetrics.currentLatency < defaultLatency.max) {
         userMetrics.currentLatency += defaultLatency.user
     }