+++ vscode/src/completions/tree-sitter/parser.ts
@@ -3,7 +3,7 @@ import path from 'path'
 import Parser from 'web-tree-sitter'
 
 import { SupportedLanguage } from './grammars'
-import { initQueries } from './queries'
+import { initQueries } from './query-sdk'
 
 // eslint-disable-next-line @typescript-eslint/no-require-imports,@typescript-eslint/no-var-requires
 const ParserImpl = require('web-tree-sitter') as typeof Parser