+++ vscode/src/configuration.ts
@@ -36,6 +36,15 @@ export function getConfiguration(config: ConfigGetter = vscode.workspace.getConf
         debugRegex = new RegExp('.*')
     }
 
+    let autocompleteAdvancedProvider: Configuration['autocompleteAdvancedProvider'] = config.get(
+        CONFIG_KEY.autocompleteAdvancedProvider,
+        null
+    )
+    // Backwards compatibility for old config values
+    if (autocompleteAdvancedProvider === 'unstable-fireworks') {
+        autocompleteAdvancedProvider = 'fireworks'
+    }
+
     return {
         // NOTE: serverEndpoint is now stored in Local Storage instead but we will still keep supporting the one in confg
         // to use as fallback for users who do not have access to local storage
@@ -62,7 +71,7 @@ export function getConfiguration(config: ConfigGetter = vscode.workspace.getConf
         experimentalLocalSymbols: config.get(CONFIG_KEY.experimentalLocalSymbols, false),
         experimentalCommandLenses: config.get(CONFIG_KEY.experimentalCommandLenses, false),
         experimentalEditorTitleCommandIcon: config.get(CONFIG_KEY.experimentalEditorTitleCommandIcon, false),
-        autocompleteAdvancedProvider: config.get(CONFIG_KEY.autocompleteAdvancedProvider, null),
+        autocompleteAdvancedProvider,
         autocompleteAdvancedServerEndpoint: config.get<string | null>(
             CONFIG_KEY.autocompleteAdvancedServerEndpoint,
             null